requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
jellyfish>=1.0.0
pymongo[srv,tls]>=4.0.0
python-dotenv>=1.0.0
certifi>=2023.7.22